beget/server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==43.0.3
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
lxml==6.0.0
multidict==6.6.3
numpy==2.3.1
oauth2client==4.1.3
pandas==2.3.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDrive2==1.21.3
pyOpenSSL==24.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
